src/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
google-generativeai>=0.8.0
joblib
scikit-learn
xgboost
numpy
pandas
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
asyncpg>=0.29.0
bcrypt==4.0.1
